+++ v2/src/dar_backup/command_runner.py
@@ -14,7 +14,6 @@ except ImportError:
     termios = None
 import tempfile
 import time
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 from typing import List, Optional, Union
 from dar_backup.util import get_logger